Move remaining index.js code to server.js
